requirements.txt
# GUI Framework
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Instagram API (FREE - No API Key Required!)
instagrapi==2.1.2

# Scheduling
APScheduler==3.10.4

# Image Processing (FREE)
requests==2.31.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7

# Database
# sqlite3 is built-in to Python

# Logging and Configuration
pyyaml==6.0.1
